boot, bootloader, o/snapstate: boot env manip goes in bootloader
This change moves almost everything that looks at boot environment
variables to bootloader itself. The exception (for now at least) is
`image`, which for now is special; thus the env var bits still need to
be exported for now.
