+++ boot/export_test.go
@@ -23,10 +23,6 @@ import (
 	"github.com/snapcore/snapd/snap"
 )
 
-var (
-	NameAndRevnoFromSnap = nameAndRevnoFromSnap
-)
-
 func NewCoreBootParticipant(s snap.PlaceInfo, t snap.Type) *coreBootParticipant {
 	return &coreBootParticipant{s: s, t: t}
 }
