+++ boot/kernel_os_test.go
@@ -90,128 +90,6 @@ func (s *coreBootSetSuite) TestChangeRequiresRebootError(c *C) {
 	c.Check(logbuf.String(), testutil.Contains, `cannot get boot settings: brkn`)
 }
 
-func (s *coreBootSetSuite) TestSetNextBootError(c *C) {
-	s.loader.GetErr = errors.New("zap")
-	err := boot.NewCoreBootParticipant(&snap.Info{}, snap.TypeApp).SetNextBoot()
-	c.Check(err, ErrorMatches, `cannot set next boot: zap`)
-
-	bootloader.ForceError(errors.New("brkn"))
-	err = boot.NewCoreBootParticipant(&snap.Info{}, snap.TypeApp).SetNextBoot()
-	c.Check(err, ErrorMatches, `cannot set next boot: brkn`)
-}
-
-func (s *coreBootSetSuite) TestSetNextBootForCore(c *C) {
-	info := &snap.Info{}
-	info.SnapType = snap.TypeOS
-	info.RealName = "core"
-	info.Revision = snap.R(100)
-
-	bs := boot.NewCoreBootParticipant(info, info.GetType())
-	err := bs.SetNextBoot()
-	c.Assert(err, IsNil)
-
-	v, err := s.loader.GetBootVars("snap_try_core", "snap_mode")
-	c.Assert(err, IsNil)
-	c.Assert(v, DeepEquals, map[string]string{
-		"snap_try_core": "core_100.snap",
-		"snap_mode":     "try",
-	})
-
-	c.Check(bs.ChangeRequiresReboot(), Equals, true)
-}
-
-func (s *coreBootSetSuite) TestSetNextBootWithBaseForCore(c *C) {
-	info := &snap.Info{}
-	info.SnapType = snap.TypeBase
-	info.RealName = "core18"
-	info.Revision = snap.R(1818)
-
-	bs := boot.NewCoreBootParticipant(info, info.GetType())
-	err := bs.SetNextBoot()
-	c.Assert(err, IsNil)
-
-	v, err := s.loader.GetBootVars("snap_try_core", "snap_mode")
-	c.Assert(err, IsNil)
-	c.Assert(v, DeepEquals, map[string]string{
-		"snap_try_core": "core18_1818.snap",
-		"snap_mode":     "try",
-	})
-
-	c.Check(bs.ChangeRequiresReboot(), Equals, true)
-}
-
-func (s *coreBootSetSuite) TestSetNextBootForKernel(c *C) {
-	info := &snap.Info{}
-	info.SnapType = snap.TypeKernel
-	info.RealName = "krnl"
-	info.Revision = snap.R(42)
-
-	bp := boot.NewCoreKernel(info)
-	err := bp.SetNextBoot()
-	c.Assert(err, IsNil)
-
-	v, err := s.loader.GetBootVars("snap_try_kernel", "snap_mode")
-	c.Assert(err, IsNil)
-	c.Assert(v, DeepEquals, map[string]string{
-		"snap_try_kernel": "krnl_42.snap",
-		"snap_mode":       "try",
-	})
-
-	bootVars := map[string]string{
-		"snap_kernel":     "krnl_40.snap",
-		"snap_try_kernel": "krnl_42.snap"}
-	s.loader.SetBootVars(bootVars)
-	c.Check(bp.ChangeRequiresReboot(), Equals, true)
-
-	// simulate good boot
-	bootVars = map[string]string{"snap_kernel": "krnl_42.snap"}
-	s.loader.SetBootVars(bootVars)
-	c.Check(bp.ChangeRequiresReboot(), Equals, false)
-}
-
-func (s *coreBootSetSuite) TestSetNextBootForKernelForTheSameKernel(c *C) {
-	info := &snap.Info{}
-	info.SnapType = snap.TypeKernel
-	info.RealName = "krnl"
-	info.Revision = snap.R(40)
-
-	bootVars := map[string]string{"snap_kernel": "krnl_40.snap"}
-	s.loader.SetBootVars(bootVars)
-
-	err := boot.NewCoreKernel(info).SetNextBoot()
-	c.Assert(err, IsNil)
-
-	v, err := s.loader.GetBootVars("snap_kernel")
-	c.Assert(err, IsNil)
-	c.Assert(v, DeepEquals, map[string]string{
-		"snap_kernel": "krnl_40.snap",
-	})
-}
-
-func (s *coreBootSetSuite) TestSetNextBootForKernelForTheSameKernelTryMode(c *C) {
-	info := &snap.Info{}
-	info.SnapType = snap.TypeKernel
-	info.RealName = "krnl"
-	info.Revision = snap.R(40)
-
-	bootVars := map[string]string{
-		"snap_kernel":     "krnl_40.snap",
-		"snap_try_kernel": "krnl_99.snap",
-		"snap_mode":       "try"}
-	s.loader.SetBootVars(bootVars)
-
-	err := boot.NewCoreKernel(info).SetNextBoot()
-	c.Assert(err, IsNil)
-
-	v, err := s.loader.GetBootVars("snap_kernel", "snap_try_kernel", "snap_mode")
-	c.Assert(err, IsNil)
-	c.Assert(v, DeepEquals, map[string]string{
-		"snap_kernel":     "krnl_40.snap",
-		"snap_try_kernel": "",
-		"snap_mode":       "",
-	})
-}
-
 // ubootBootSetSuite tests the uboot specific code in the bootloader handling
 type ubootBootSetSuite struct {
 	baseBootSetSuite
